# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Profile server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/movies": {
            "get": {
                "description": "Get All Movie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "summary": "Movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Movie",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Movie",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit Movie",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort Movie",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.GetAllMovie"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Movies",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Edit tittle",
                        "name": "tittle",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update genre",
                        "name": "genre",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Update Image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Synopsis",
                        "name": "synopsis",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Author",
                        "name": "author",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Actors",
                        "name": "actors",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Realease Date",
                        "name": "release_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Duration",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Tag",
                        "name": "tag",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.MoviesbyTag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "get": {
                "description": "Detail Movies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Detail Movie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.MoviesNoTag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Movies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete Movie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.MoviesbyTag"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit Movies",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movies"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Id Movie",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Edit tittle",
                        "name": "tittle",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update genre",
                        "name": "genre",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Update Image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Synopsis",
                        "name": "synopsis",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Author",
                        "name": "author",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Actors",
                        "name": "actors",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Realease Date",
                        "name": "release_date",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Duration",
                        "name": "duration",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Tag",
                        "name": "tag",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Movie_body"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "page_info": {},
                "results": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.GetAllMovie": {
            "type": "object",
            "properties": {
                "genre": {
                    "type": "string",
                    "example": "Action"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "Spiderman.jpg"
                },
                "tittle": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
        "models.Movie_body": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "genre": {
                    "type": "string",
                    "example": "Action"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "Spiderman.jpg"
                },
                "release_date": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string",
                    "example": "film action universal"
                },
                "tag": {
                    "description": "Release_date time.Time `json:\"release_date\" form:\"release_date\"`\nDuration     time.Time `json:\"duration\" form:\"duration\"`",
                    "type": "string"
                },
                "tittle": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
...
        "models.MoviesNoTag": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "duration": {
                    "type": "string"
                },
                "genre": {
                    "type": "string",
                    "example": "Action"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "Spiderman.jpg"
                },
                "release_date": {
                    "type": "string"
                },
                "synopsis": {
                    "type": "string",
                    "example": "film action universal"
                },
                "tittle": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
        "models.MoviesbyTag": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "genre": {
                    "type": "string",
                    "example": "Action"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "Spiderman.jpg"
                },
                "synopsis": {
                    "type": "string",
                    "example": "film action universal"
                },
                "tag": {
                    "description": "Release_date time.Time `json:\"release_date\" form:\"release_date\"`\nDuration     time.Time `json:\"duration\" form:\"duration\"`",
                    "type": "string"
                },
                "tittle": {
                    "type": "string",
                    "example": "Spiderman"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}